split FamilyList component into two components (create Family component)
